chore: drop obsolete previous ddb module
