+++ bench-vortex/src/bin/clickbench.rs
@@ -68,47 +68,6 @@ struct Args {
     hide_progress_bar: bool,
     #[arg(long, default_value_t = false)]
     show_metrics: bool,
-    #[arg(long)]
-    skip_duckdb_build: bool,
-}
-
-struct DataFusionCtx {
-    execution_plans: Vec<(usize, Arc<dyn ExecutionPlan>)>,
-    metrics: Vec<(
-        usize,
-        Format,
-        Vec<datafusion::physical_plan::metrics::MetricsSet>,
-    )>,
-
-    session: SessionContext,
-    emit_plan: bool,
-}
-
-enum EngineCtx {
-    DataFusion(DataFusionCtx),
-    DuckDB(ddb2::DuckDBCtx),
-}
-
-impl EngineCtx {
-    fn new_with_datafusion(session_ctx: SessionContext, emit_plan: bool) -> Self {
-        EngineCtx::DataFusion(DataFusionCtx {
-            execution_plans: Vec::new(),
-            metrics: Vec::new(),
-            session: session_ctx,
-            emit_plan,
-        })
-    }
-
-    fn new_with_duckdb() -> anyhow::Result<Self> {
-        Ok(EngineCtx::DuckDB(ddb2::DuckDBCtx::new()?))
-    }
-
-    fn to_engine(&self) -> Engine {
-        match &self {
-            EngineCtx::DuckDB(_) => Engine::DuckDB,
-            EngineCtx::DataFusion(_) => Engine::DataFusion,
-        }
-    }
 }
 
 fn main() -> anyhow::Result<()> {
